+++ R/cff-class.R
@@ -1,0 +1,10 @@
+#' The `cff` class
+#'
+#'
+#' @name cff-class
+
+#' @description
+#' TODO
+#'
+#'
+NULL